Sort by text-only representation of the user, not HTML
/Elements/ShowUser always includes a <span data-user-id="..."> element;
thus, sorting the result leads to sorting ASCII-betically by id -- this
is unhelpful, to say the least.

Switch to sorting by $user->Format; this is, strictly, not guaranteed to
be the same as the ShowUser component's text, as the latter has a
callback that might modify it.  However, it keeps compatibility with
this callsite respecting said callback, and still wraps in data-user-id
for metadata (currently unused), while sorting by the presumed display.

The computational overhead to effectively calling ->Format twice per
user is judged to be small; should that cease being the case, it could
be cached on the user object.

Fixes: I#30771