Improve hiding/showing of rights UI for unselected principal
If you're assigning rights to a user or group and you're using the "Add
Group" or "Add User" text field but have not yet typed in a valid group
or user name, RT will still display the three-tab list of checkboxes UI
on the right-hand side of the page. This is confusing to administrators
since they should not be selecting rights, as there is no principal
selected yet, but we show the UI anyway.

This re-engineering hides and shows the panel itself based on whether
there is a valid principal selected. This also improves the
responsiveness of the "no user/group selected" warning which previously
required you to blur the autocomplete field to render. Now, it provides
more instantaneous feedback.